+++ cmd/fyne/commands/build.go
@@ -5,11 +5,13 @@ import (
 	"os"
 	"os/exec"
 	"runtime"
+	"strings"
 )
 
 type builder struct {
 	os, srcdir string
 	release    bool
+	tags       []string
 }
 
 func (b *builder) build() error {
@@ -18,24 +20,37 @@ func (b *builder) build() error {
 		goos = targetOS()
 	}
 
-	var cmd *exec.Cmd
-	if goos == "windows" {
-		if b.release {
-			cmd = exec.Command("go", "build", "-ldflags", "-s -w -H=windowsgui", ".")
-		} else {
-			cmd = exec.Command("go", "build", "-ldflags", "-H=windowsgui", ".")
-		}
-	} else {
-		if b.release {
-			cmd = exec.Command("go", "build", "-ldflags", "-s -w", ".")
-		} else {
-			cmd = exec.Command("go", "build", ".")
-		}
-	}
-	cmd.Dir = b.srcdir
+	args := []string{"build"}
 	env := os.Environ()
 	env = append(env, "CGO_ENABLED=1") // in case someone is trying to cross-compile...
 
+	if goos == "windows" {
+		if b.release {
+			args = append(args, "-ldflags", "-s -w -H=windowsgui")
+		} else {
+			args = append(args, "-ldflags", "-H=windowsgui")
+		}
+	} else {
+		if goos == "darwin" {
+			env = append(env, "CGO_CFLAGS=-mmacosx-version-min=10.11")
+			env = append(env, "CGO_LDFLAGS=-mmacosx-version-min=10.11")
+		}
+		if b.release {
+			args = append(args, "-ldflags", "-s -w")
+		}
+	}
+
+	// handle build tags
+	tags := b.tags
+	if b.release {
+		tags = append(tags, "release")
+	}
+	if len(tags) > 0 {
+		args = append(args, "-tags", strings.Join(tags, ","))
+	}
+
+	cmd := exec.Command("go", args...)
+	cmd.Dir = b.srcdir
 	if goos != "ios" && goos != "android" {
 		env = append(env, "GOOS="+goos)
 	}