+++ dialog/color.go
@@ -14,18 +14,22 @@ import (
 )
 
 // ColorPickerDialog is a simple dialog window that displays a color picker.
+//
+// Since: 1.4
 type ColorPickerDialog struct {
 	*dialog
 	Advanced bool
 	color    color.Color
 	callback func(c color.Color)
-	advanced *widget.AccordionContainer
+	advanced *widget.Accordion
 	picker   *colorAdvancedPicker
 }
 
 // NewColorPicker creates a color dialog and returns the handle.
 // Using the returned type you should call Show() and then set its color through SetColor().
 // The callback is triggered when the user selects a color.
+//
+// Since: 1.4
 func NewColorPicker(title, message string, callback func(c color.Color), parent fyne.Window) *ColorPickerDialog {
 	p := &ColorPickerDialog{
 		dialog:   newDialog(title, message, theme.ColorPaletteIcon(), nil /*cancel?*/, parent),
@@ -37,6 +41,8 @@ func NewColorPicker(title, message string, callback func(c color.Color), parent 
 
 // ShowColorPicker creates and shows a color dialog.
 // The callback is triggered when the user selects a color.
+//
+// Since: 1.4
 func ShowColorPicker(title, message string, callback func(c color.Color), parent fyne.Window) {
 	NewColorPicker(title, message, callback, parent).Show()
 }
@@ -88,7 +94,7 @@ func (p *ColorPickerDialog) updateUI() {
 		p.picker = newColorAdvancedPicker(p.color, func(c color.Color) {
 			p.color = c
 		})
-		p.advanced = widget.NewAccordionContainer(widget.NewAccordionItem("Advanced", p.picker))
+		p.advanced = widget.NewAccordion(widget.NewAccordionItem("Advanced", p.picker))
 
 		p.dialog.content = fyne.NewContainerWithLayout(layout.NewVBoxLayout(),
 			fyne.NewContainerWithLayout(layout.NewCenterLayout(),