+++ widget/radio_group.go
@@ -8,6 +8,8 @@ import (
 
 // RadioGroup widget has a list of text labels and radio check icons next to each.
 // Changing the selection (only one can be selected) will trigger the changed func.
+//
+// Since: 1.4
 type RadioGroup struct {
 	DisableableWidget
 	Horizontal bool
@@ -22,6 +24,8 @@ type RadioGroup struct {
 var _ fyne.Widget = (*RadioGroup)(nil)
 
 // NewRadioGroup creates a new radio widget with the set options and change handler
+//
+// Since: 1.4
 func NewRadioGroup(options []string, changed func(string)) *RadioGroup {
 	r := &RadioGroup{
 		DisableableWidget: DisableableWidget{},